requirements.txt
fastapi
uvicorn
requests
pydantic>=2.8.0
pytest
httpx
python-dotenv
pytest-cov
